+++ examples/searchcommands_app/package/bin/sum.py
@@ -17,9 +17,11 @@
 
 from __future__ import absolute_import, division, print_function, unicode_literals
 import app
+import os,sys
 
+splunkhome = os.environ['SPLUNK_HOME']
+sys.path.append(os.path.join(splunkhome, 'etc', 'apps', 'searchcommands_app', 'lib'))
 from splunklib.searchcommands import dispatch, ReportingCommand, Configuration, Option, validators
-import sys
 
 
 @Configuration(requires_preop=True)